ECP-28: build color repaired

+++ tools/build.sh
@@ -23,15 +23,22 @@ remove_directory(){
 }
 remove_files(){
 	DIRECTORY=$1
-	FILE=$2
-	find $DIRECTORY -name $FILE -nowarn -exec rm -rf {} \;
-	echo "- Delete $FILE : ""$VERT""DONE""$NORMAL"""
+        FILE=$2
+        if [ -f "${DIRECTORY}/${FILE}" ]
+        then
+            find $DIRECTORY -name $FILE -nowarn -exec rm -rf {} \;
+            echo -e "- Delete ${FILE} : ${VERT}DONE${NORMAL}"
+        fi
+        if [ -d "${DIRECTORY}/${FILE}" ]
+        then
+            rm -Rf ${DIRECTORY}/${FILE}
+        fi
 }
 
 remove_directories(){
 	DIRECTORY=$1
 	find $DIRECTORY -maxdepth 1 -mindepth 1 -type d -exec rm -rf {} \;
-	echo "- Delete $FILE : ""$VERT""DONE""$NORMAL"""
+	echo -e "- Delete $FILE : ${VERT}DONE${NORMAL}"
 }
 # Check parameters
 if [ -z "$1" ]; then
@@ -51,33 +58,37 @@ FOLDER_EXPORT="/tmp/lengow-woocommerce/export"
 FOLDER_TOOLS="/tmp/lengow-woocommerce/tools"
 FOLDER_TRANSLATION="/tmp/lengow-woocommerce/translations/yml"
 
-VERT="\\033[1;32m"
-ROUGE="\\033[1;31m"
-NORMAL="\\033[0;39m"
-BLEU="\\033[1;36m"
+VERT="\e[32m"
+ROUGE="\e[31m"
+NORMAL="\e[39m"
+BLEU="\e[36m"
+
 
 # Process
 echo
 echo "#####################################################"
 echo "##                                                 ##"
-echo "##       ""$BLEU""Lengow Woocommerce""$NORMAL"" - Build Module          ##"
+echo -e "##       "${BLEU}Lengow Magento${NORMAL}" - Build Module             ##"
 echo "##                                                 ##"
 echo "#####################################################"
 echo
-FOLDER="$(dirname "$(pwd)")"
+PWD=$(pwd)
+FOLDER=$(dirname ${PWD})
 echo $FOLDER
 if [ ! -d "$FOLDER" ]; then
-	echo "Folder doesn't exist : ""$ROUGE""ERROR""$NORMAL"""
+	echo -e "Folder doesn't exist : ${ROUGE}ERROR${NORMAL}"
 	echo
 	exit 0
 fi
+PHP=$(which php8.1)
+echo ${PHP}
 
 # Generate translations
-php translate.php
-echo "- Generate translations : ""$VERT""DONE""$NORMAL"""
+${PHP} translate.php
+echo -e "- Generate translations : ${VERT}DONE${NORMAL}"
 # Create files checksum
-php checkmd5.php
-echo "- Create files checksum : ""$VERT""DONE""$NORMAL"""
+${PHP} checkmd5.php
+echo -e "- Create files checksum : ${VERT}DONE${NORMAL}"
 #remove TMP FOLDER
 remove_directory $FOLDER_TMP
 #copy files
@@ -102,25 +113,25 @@ remove_files $FOLDER_TMP "Jenkinsfile"
 remove_files $FOLDER_CONFIG "marketplaces.json"
 # Clean Log Folder
 remove_files $FOLDER_LOGS "*.txt"
-echo "- Clean logs folder : ""$VERT""DONE""$NORMAL"""
+echo -e "- Clean logs folder : ${VERT}DONE${NORMAL}"
 # Clean export folder
 remove_files $FOLDER_EXPORT "*.csv"
 remove_files $FOLDER_EXPORT "*.yaml"
 remove_files $FOLDER_EXPORT "*.json"
 remove_files $FOLDER_EXPORT "*.xml"
-echo "- Clean export folder : ""$VERT""DONE""$NORMAL"""
+echo -e "- Clean export folder : ${VERT}DONE${NORMAL}"
 # Clean export folder
 remove_directory $FOLDER_TOOLS
-echo "- Remove Tools folder : ""$VERT""DONE""$NORMAL"""
+echo -e "- Remove Tools folder : ${VERT}DONE${NORMAL}"
 #remove TMP FOLDER_TRANSLATION
 remove_directory $FOLDER_TRANSLATION
-echo "- Remove Translation yml folder : ""$VERT""DONE""$NORMAL"""
+echo -e "- Remove Translation yml folder : ${VERT}DONE${NORMAL}"
 
 # Make zip
 cd /tmp
-zip "-r" $ARCHIVE_NAME "lengow-woocommerce"
-echo "- Build archive : ""$VERT""DONE""$NORMAL"""
-if [ -d  ~/Bureau ]
+zip -r $ARCHIVE_NAME "lengow-woocommerce"
+echo -e "- Build archive : ${VERT}DONE${NORMAL}"
+if [ -d  "~/Bureau" ]
 then
     mv $ARCHIVE_NAME ~/Bureau
 else
